requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-bigquery>=3.14.0
python-dotenv==1.0.0
